refactor: ISession type used in controllers
